Slim down the tutorials
Now that all of the packages have docs with tests, and we have a whole lot more testing hardware than we ever had before, we are starting to slim down the tutorials (which only get rebuilt and fixed periodically) and moving them to the individual package docs. We are doing this and turning all of the docs into downstream tests (https://github.com/SciML/SciMLDocs/issues/104) so that they are perpetually working. It will be a massive amount of compute, but we have that now, so let's make everything super robust.

Related PRs:

- https://github.com/SciML/JumpProcesses.jl/pull/277
- https://github.com/SciML/DiffEqParamEstim.jl/pull/188
- https://github.com/SciML/DiffEqBayes.jl/pull/274